HTTPS impersonation working with LittleProxy
It's derived from the OWASP Zed Attack Proxy Project, which is licensed
under the Apache License, Version 2.0. The certificate authority is
created lazily with Bouncy Castle. The key store type is PKCS12, which
has to be implemented by every vendor. The server certificates won't be
cached, but the connection is already cached. It lacks upstream
certificate validation(!) and a proper exception handling.