Added the AcceptHandler feature to the proxy.
Using the withAcceptHandler() method on the proxy bootstrap is now
possible to get access to the byte stream of the first read performed
on the underlying channel after a connection has been accepted and
before the message is http-parsed inside the channel pipeline.

An implementation of the interface AcceptHandler can be able to read
and/or modify the first bytes sent by a client before they are parsed,
as well as close the connection if some requirements are not met
and so on.